[DSCP-467] Reorganize modules
