+++ witness/src/Dscp/Web/Swagger/Generation.hs
@@ -1,0 +1,2 @@
+-- | Means of automatic swagger documentation generation.
+{-# OPTIONS_GHC -F -pgmF autoexporter -Wno-dodgy-exports -Wno-unused-imports #-}
