+++ witness/src/Dscp/Web/Swagger/Instances.hs
@@ -1,0 +1,1 @@
+{-# OPTIONS_GHC -F -pgmF autoexporter -Wno-dodgy-exports -Wno-unused-imports #-}
